getting ready for vercel deployment
