+++ package.json
@@ -51,6 +51,7 @@
     "next": "14.2.16",
     "next-themes": "^0.4.4",
     "react": "^18",
+    "react-day-picker": "^8.7.1",
     "react-dom": "^18",
     "react-error-boundary": "^5.0.0",
     "react-hook-form": "^7.54.1",
